Generate metadata for config props
Adds the `spring-boot-configuration-processor` to enable automatic metadata
generation so our config props play well in IDEs.

Also updates the default values for duration properties to use
`Duration.of<Unit>(long)` instead of `Duration.of(long, DurationUnit)` as
Spring Boot config processor can not determine a default for the config
metadata in the latter format.